Started Entity Implementation on View
Little Refactoring

Changed Double Coords. to Integer Coords

+++ src/main/java/input/api/ViewEventListener.java
@@ -4,6 +4,6 @@ import model.impl.Pair;
 
 public interface ViewEventListener {
 
-    void notifyMouseEvent(Pair<Double, Double> pos);
+    void notifyMouseEvent(Pair<Integer, Integer> pos);
 
 }
